synthesizer/requirements.txt
paho-mqtt
numpy
pandas
python-dotenv
